+++ ECourseWeb/src/main/java/com/htt/repository/LessonRepository.java
@@ -16,7 +16,7 @@ public interface LessonRepository {
     List<Lesson> getLessons(Map<String, String> params);
     List<Lesson> getLessons();
     void addOrUpdate(Lesson c);
-    Lesson getLessonById(int id);
-    void deleteLesson(int id);
+    Lesson getLessonById(Long id);
+    void deleteLesson(Long id);
     List<Lesson> getLessonsByCourseId(Long courseId);
 }
